The player can now only kill zombies while driving

+++ Assets/Scripts/Killzone.cs
@@ -2,11 +2,23 @@
 
 public class Killzone : MonoBehaviour
 {
-    private void OnTriggerEnter(Collider other)
+    [SerializeField] private bool _front;
+    [SerializeField] private Car _player;
+
+    private void OnTriggerStay(Collider other)
     {
         if (!other.CompareTag("Enemy")) return;
         if (other.GetComponent<Zombie>().IsDead()) return;
-        
-        ZombieManager.Instance.Kill(other.transform);
+
+        if (_front && _player.GetPower() > 0)
+        {
+            ZombieManager.Instance.Kill(other.transform);
+            return;
+        }
+
+        if (!_front && _player.GetPower() < 0)
+        {
+            ZombieManager.Instance.Kill(other.transform);
+        }
     }
 }
